+++ core/src/com/mygdx/game/entities/Spring.java
@@ -37,8 +37,15 @@ public class Spring extends Entity {
     public void render(SpriteBatch batch) {
         if (appear) {
             batch.draw(getCurrentFrame(),
-                    (int) this.x, (int) this.y,
+//                    x * handler.getWorld_to_scene_width(),
+//                    y * handler.getWorld_to_scene_height(),
+//                    width * handler.getWorld_to_scene_width(),
+//                    height * handler.getWorld_to_scene_height());
+
+                     this.x,  this.y,
                     this.width, this.height);
+
+//                    getSceneWidth(), getSceneHeight());
         }
     }
